Move App to components folder
