+++ src/components/App/App.js
@@ -1,0 +1,14 @@
+import Header from '../Header/Header';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <Header />
+      </header>
+    </div>
+  );
+}
+
+export default App;
